Still Picture - Automatic Email as MIME Attachment
The app now has the ability to take a still picture and automatically
email that picture as a MIME attachment without the user knowing.  This
functionality required custom built-in email sub-routines, because the
app has to send digital data  to our servers seamlessly.  Besides, we
don't want to store any data acquired on the data collector's device.